requirements.txt
numpy
pyqt5
pillow
gym==0.21.0
gym-retro==0.8.0
